refactor: 엔티티와 파일 리펙토링
- entity refactoring
- 프로젝트 구조 변경
