+++ src/main/java/com/genius/todoffin/hits/repository/HitsRepository.java
@@ -1,6 +1,6 @@
-package com.genius.todoffin.challenge.repository;
+package com.genius.todoffin.hits.repository;
 
-import com.genius.todoffin.challenge.domain.Hits;
+import com.genius.todoffin.hits.domain.Hits;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface HitsRepository extends JpaRepository<Hits, Long> {
