+++ src/main/java/com/genius/todoffin/instance/repository/InstanceRepository.java
@@ -1,6 +1,6 @@
-package com.genius.todoffin.challenge.repository;
+package com.genius.todoffin.instance.repository;
 
-import com.genius.todoffin.challenge.domain.Instance;
+import com.genius.todoffin.instance.domain.Instance;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InstanceRepository extends JpaRepository<Instance, Long> {
