+++ src/main/java/com/genius/todoffin/participantinfo/domain/JoinStatus.java
@@ -1,0 +1,5 @@
+package com.genius.todoffin.participantinfo.domain;
+
+public enum JoinStatus {
+    NO, YES
+}
