+++ src/main/java/com/genius/todoffin/topic/repository/TopicRepository.java
@@ -1,6 +1,6 @@
-package com.genius.todoffin.challenge.repository;
+package com.genius.todoffin.topic.repository;
 
-import com.genius.todoffin.challenge.domain.Topic;
+import com.genius.todoffin.topic.domain.Topic;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TopicRepository extends JpaRepository<Topic, Long> {
